// client/src/mutations/__generated__/CreateProject.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: CreateProject
// ====================================================

export interface CreateProject {
  createProject: string;
}

export interface CreateProjectVariables {
  id: string;
  category: string;
  title: string;
  description: string;
  url: string;
  goal: string;
  deadline: string;
}