// client/src/mutations/__generated__/EditProject.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: EditProject
// ====================================================

export interface EditProject {
  editProject: string;
}

export interface EditProjectVariables {
  address: string;
  category: string;
  title: string;
  description: string;
  url: string;
  goal: string;
  deadline: string;
}