// client/src/queries/__generated__/GetProjectAddress.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: GetProjectAddress
// ====================================================

export interface GetProjectAddress {
  projectAddress: string | null;
}

export interface GetProjectAddressVariables {
  index: string;
}