// client/src/queries/__generated__/GetProjectIndex.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: GetProjectIndex
// ====================================================

export interface GetProjectIndex {
  projectIndex: string;
}

export interface GetProjectIndexVariables {
  id: string;
}